text_processing/requirements.txt
wagtail-whoosh